Refactored components and lifted state up for user to App component
