+++ client/src/components/layout/Dashboard.js
@@ -1,50 +1,27 @@
 import React, { useState, useEffect } from "react";
 import { Redirect } from "react-router-dom";
-import { getUserData } from "../../actions/authActions";
 import isEmpty from "is-empty";
-import cookies from "js-cookie";
 
-import Navbar from "./Navbar";
-import DashboardBody from "./DashboardBody";
+import Typography from '@material-ui/core/Grid';
 
-function Dashboard() {
-  const token = cookies.get("user");
-
-  const [isLoggedIn, setLoggedIn] = useState(!isEmpty(token));
-  const [user, setUser] = useState({
-		firstName: "",
-		lastName: "",
-	});
+function Dashboard(props) {
+  const [isLoggedIn, setLoggedIn] = useState(props.loggedIn);
+  const [user, setUser] = useState(props.userData);
 	
   useEffect(() => {
-    if (isLoggedIn) {
-      const getUserPayload = async () => {
-        const data = await getUserData(token);
-				const userPayload = await data.payload;
-				
-        setUser({
-					firstName: userPayload.firstName,
-					lastName: userPayload.lastName
-        });
-      }
+    setUser(props.userData)
+	}, [props.userData]);
 
-      getUserPayload();
-    }
-	}, [isLoggedIn, token]);
+	useEffect(() => {
+		setLoggedIn(props.loggedIn);
+	}, [props.loggedIn]);
 	
-	const handleLoggedIn = async stillLoggedIn => {
-		if (stillLoggedIn) {
-			setLoggedIn(true);
-		} else {
-			setLoggedIn(false);
-		}
-	}
-
 	return (
 		!isLoggedIn ? <Redirect to="/login" /> : 
 		<div>
-			<Navbar changeLoggedIn={handleLoggedIn} isLoggedIn={isLoggedIn} />
-			<DashboardBody userData={user}  />
+			<Typography>
+				Welcome {!isEmpty(user.firstName) && user.firstName}
+			</Typography>
 		</div>
 	);
 }
