+++ client/src/components/layout/Landing.js
@@ -1,13 +1,29 @@
 import React from "react";
+import Register from "../auth/Register";
 
-import Navbar from "./Navbar";
-import LandingBody from "./LandingBody";
+import { makeStyles } from '@material-ui/core/styles';
+import { Grid } from '@material-ui/core';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    flexGrow: 1,
+  },
+  position: {
+		padding: '125px 0px 100px 20px',
+		marginLeft: '0px'
+  },
+}));
 
 function Landing() {
+	const classes = useStyles();
+	
 	return (
-    <div>
-      <Navbar />
-      <LandingBody />
+    <div className="landing-background background-image-full">
+			<div className={classes.root}>
+				<Grid className={classes.position} item xs={4}>
+					<Register />
+				</Grid>
+  	  </div>
     </div>
 	);
 }
